+++ packages/types-core/src/global.ts
@@ -2,13 +2,14 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import Pino from 'pino';
-import {DynamicDatasourceCreator} from './interfaces';
+import {Cache, DynamicDatasourceCreator} from './interfaces';
 import {Store} from './store';
 
 // base global
 declare global {
   const logger: Pino.Logger;
   const store: Store;
+  const cache: Cache;
   const chainId: string;
   const createDynamicDatasource: DynamicDatasourceCreator;
 }