+++ packages/types-core/src/interfaces.ts
@@ -2,3 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0
 
 export type DynamicDatasourceCreator = (name: string, args: Record<string, unknown>) => Promise<void>;
+
+export interface Cache {
+  set<D = string>(key: string, value: D): Promise<void>;
+  get<D = string>(key: string): Promise<D | undefined>;
+}